// src/core/renderer/interfaces/renderer.interface.ts
import type { TTypewriterState } from "../../state/types/typewriter-state.type";



/**
 * @description
 * Interface for all typewriter renderers.
 * A renderer receives state and is responsible for displaying it.
 */
export interface IRenderer {
  render: (state: TTypewriterState) => void | Promise<void>;
  mount?: (state: TTypewriterState) => void | Promise<void>;
  unmount?: () => void | Promise<void>;
}